+++ pretrained/tokenizer_test.go
@@ -2,6 +2,7 @@ package pretrained
 
 import (
 	"log"
+	"reflect"
 	"testing"
 
 	"github.com/sugarme/tokenizer"
@@ -172,3 +173,176 @@ func TestFromFile(t *testing.T) {
 }
 
 */
+
+// Test Bert model `bert-base-uncased`
+func TestBert(t *testing.T) {
+	configFile, err := tokenizer.CachedPath("bert-base-uncased", "tokenizer.json")
+	if err != nil {
+		panic(err)
+	}
+
+	tk, err := FromFile(configFile)
+	if err != nil {
+		panic(err)
+	}
+
+	sentence := `The Gophers craft code using [MASK] language.`
+	en, err := tk.EncodeSingle(sentence, true)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	gotTokens := en.Tokens
+	wantTokens := []string{
+		"[CLS]", "the", "go", "##pher", "##s", "craft", "code", "using", "[MASK]", "language", ".", "[SEP]",
+	}
+	if !reflect.DeepEqual(wantTokens, gotTokens) {
+		t.Errorf("want %v,\ngot %v\n", wantTokens, gotTokens)
+	}
+
+	gotIds := en.Ids
+	wantIds := []int{101, 1996, 2175, 27921, 2015, 7477, 3642, 2478, 103, 2653, 1012, 102}
+	if !reflect.DeepEqual(wantIds, gotIds) {
+		t.Errorf("want %v,\ngot %v\n", wantIds, gotIds)
+	}
+
+	gotOffsets := en.Offsets
+	wantOffsets := [][]int{
+		{0, 0},
+		{0, 3},
+		{4, 6},
+		{6, 10},
+		{10, 11},
+		{12, 17},
+		{18, 22},
+		{23, 28},
+		{29, 35},
+		{36, 44},
+		{44, 45},
+		{0, 0},
+	}
+	if !reflect.DeepEqual(wantOffsets, gotOffsets) {
+		t.Errorf("want %v,\ngot %v\n", wantOffsets, gotOffsets)
+	}
+}
+
+// Test Roberta model `roberta-base`
+func TestRoberta(t *testing.T) {
+	configFile, err := tokenizer.CachedPath("roberta-base", "tokenizer.json")
+	if err != nil {
+		panic(err)
+	}
+
+	tk, err := FromFile(configFile)
+	if err != nil {
+		panic(err)
+	}
+
+	sentence := `The Gophers craft code using [MASK] language.`
+	en, err := tk.EncodeSingle(sentence, true)
+	if err != nil {
+		log.Fatal(err)
+	}
+	/*
+		sequence: The Gophers craft code using [MASK] language.
+		tokens: ['<s>', 'The', 'ĠG', 'ophers', 'Ġcraft', 'Ġcode', 'Ġusing', 'Ġ[', 'MAS', 'K', ']', 'Ġlanguage', '.', '</s>']
+		ids: [0, 133, 272, 30482, 6306, 3260, 634, 646, 32804, 530, 742, 2777, 4, 2]
+		offsets: [(0, 0), (0, 3), (4, 5), (5, 11), (12, 17), (18, 22), (23, 28), (29, 30), (30, 33), (33, 34), (34, 35), (36, 44), (44, 45), (0, 0)]
+	*/
+
+	gotTokens := en.Tokens
+	wantTokens := []string{
+		"<s>", "The", "ĠG", "ophers", "Ġcraft", "Ġcode", "Ġusing", "Ġ[", "MAS", "K", "]", "Ġlanguage", ".", "</s>",
+	}
+	if !reflect.DeepEqual(wantTokens, gotTokens) {
+		t.Errorf("\nwant %q,\ngot  %q\n", wantTokens, gotTokens)
+	}
+
+	gotIds := en.Ids
+	wantIds := []int{0, 133, 272, 30482, 6306, 3260, 634, 646, 32804, 530, 742, 2777, 4, 2}
+	if !reflect.DeepEqual(wantIds, gotIds) {
+		t.Errorf("\nwant %v,\ngot  %v\n", wantIds, gotIds)
+	}
+
+	gotOffsets := en.Offsets
+	wantOffsets := [][]int{
+		{0, 0},
+		{0, 3},
+		{4, 5},
+		{5, 11},
+		{12, 17},
+		{18, 22},
+		{23, 28},
+		{29, 30},
+		{30, 33},
+		{33, 34},
+		{34, 35},
+		{36, 44},
+		{44, 45},
+		{0, 0},
+	}
+	if !reflect.DeepEqual(wantOffsets, gotOffsets) {
+		t.Errorf("\nwant %v,\ngot  %v\n", wantOffsets, gotOffsets)
+	}
+}
+
+// Test Llama model `hf-internal-testing/llama-tokenizer`
+func TestLlama(t *testing.T) {
+	configFile, err := tokenizer.CachedPath("hf-internal-testing/llama-tokenizer", "tokenizer.json")
+	if err != nil {
+		panic(err)
+	}
+
+	tk, err := FromFile(configFile)
+	if err != nil {
+		panic(err)
+	}
+
+	sentence := `The Gophers craft code using [MASK] language.`
+	en, err := tk.EncodeSingle(sentence, true)
+	if err != nil {
+		log.Fatal(err)
+	}
+	/*
+		tokens: ['<s>', '▁The', '▁G', 'oph', 'ers', '▁craft', '▁code', '▁using', '▁[', 'MA', 'SK', ']', '▁language', '.']
+		ids: [1, 450, 402, 3021, 414, 25554, 775, 773, 518, 1529, 16033, 29962, 4086, 29889]
+		offsets: [(0, 0), (0, 3), (3, 5), (5, 8), (8, 11), (11, 17), (17, 22), (22, 28), (28, 30), (30, 32), (32, 34), (34, 35), (35, 44), (44, 45)]
+	*/
+
+	gotTokens := en.Tokens
+	wantTokens := []string{
+		"<s>", "▁The", "▁G", "oph", "ers", "▁craft", "▁code", "▁using", "▁[", "MA", "SK", "]", "▁language", ".",
+	}
+	if !reflect.DeepEqual(wantTokens, gotTokens) {
+		t.Errorf("\nwant %q,\ngot  %q\n", wantTokens, gotTokens)
+	}
+
+	gotIds := en.Ids
+	wantIds := []int{
+		1, 450, 402, 3021, 414, 25554, 775, 773, 518, 1529, 16033, 29962, 4086, 29889,
+	}
+	if !reflect.DeepEqual(wantIds, gotIds) {
+		t.Errorf("\nwant %v,\ngot  %v\n", wantIds, gotIds)
+	}
+
+	gotOffsets := en.Offsets
+	wantOffsets := [][]int{
+		{0, 0},
+		{0, 3},
+		{3, 5},
+		{5, 8},
+		{8, 11},
+		{11, 17},
+		{17, 22},
+		{22, 28},
+		{28, 30},
+		{30, 32},
+		{32, 34},
+		{34, 35},
+		{35, 44},
+		{44, 45},
+	}
+	if !reflect.DeepEqual(wantOffsets, gotOffsets) {
+		t.Errorf("\nwant %v,\ngot  %v\n", wantOffsets, gotOffsets)
+	}
+}